backend/requirements.txt
boto3>=1.34.0
boto3-stubs[dynamodb,cognito-idp,bedrock-runtime,s3]>=1.34.0
pydantic>=2.5.0
python-dateutil>=2.8.2
ulid-py>=1.1.0
pytest>=7.4.0
pytest-mock>=3.12.0
moto[dynamodb,cognitoidp,s3]>=5.0.0
black>=24.0.0
ruff>=0.2.0
